slop module added
